+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [Unreleased](https://github.com/AbdelrahmanBayoumi/Azkar-App/compare/1.3.0...HEAD)
 
+### Enhancement
+
+- 🌙 Add a **System default** theme option that follows the OS light/dark setting [#113](https://github.com/AbdelrahmanBayoumi/Azkar-App/issues/113)
+
 ## [1.3.0](https://github.com/AbdelrahmanBayoumi/Azkar-App/releases/tag/1.3.0) - 2025-05-06
 
 ### Add